requirements.dev.txt
pytest
pytest-cov
requests-mock
beautifulsoup4
django-ebhealthcheck
django-health-check
selenium
apscheduler
